Split form into 3 subforms
