+++ src/pages/CreateRiddlePage.tsx
@@ -1,5 +1,5 @@
 import type { FC } from 'react';
 
-import { CreateRiddleForm } from '../components/forms/riddleForm/CreateRiddleForm.tsx';
+import { CreateRiddleForm } from '../components/forms/riddleUpsertForm/CreateRiddleForm.tsx';
 
 export const CreateRiddlePage: FC = () => <CreateRiddleForm />;
